backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
pydantic-settings
google-generativeai
PyPDF2
psycopg2-binary
